#1 feat: save/load 'account' state in Account.js
