+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\github\\budget-account-js\\src\\index.js":"1","D:\\github\\budget-account-js\\src\\App.js":"2","D:\\github\\budget-account-js\\src\\views\\Landing.js":"3","D:\\github\\budget-account-js\\src\\views\\Account.js":"4","D:\\github\\budget-account-js\\src\\views\\components\\AccountTable.js":"5"},{"size":209,"mtime":1616854620329,"results":"6","hashOfConfig":"7"},{"size":452,"mtime":1616854620328,"results":"8","hashOfConfig":"7"},{"size":748,"mtime":1616857057866,"results":"9","hashOfConfig":"7"},{"size":947,"mtime":1616942568382,"results":"10","hashOfConfig":"7"},{"size":2778,"mtime":1617536349630,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1losux8",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\github\\budget-account-js\\src\\index.js",[],"D:\\github\\budget-account-js\\src\\App.js",[],"D:\\github\\budget-account-js\\src\\views\\Landing.js",[],"D:\\github\\budget-account-js\\src\\views\\Account.js",[],"D:\\github\\budget-account-js\\src\\views\\components\\AccountTable.js",[]]
+[{"D:\\github\\budget-account-js\\src\\index.js":"1","D:\\github\\budget-account-js\\src\\App.js":"2","D:\\github\\budget-account-js\\src\\views\\Landing.js":"3","D:\\github\\budget-account-js\\src\\views\\Account.js":"4","D:\\github\\budget-account-js\\src\\views\\components\\AccountTable.js":"5"},{"size":209,"mtime":1616854620329,"results":"6","hashOfConfig":"7"},{"size":452,"mtime":1616854620328,"results":"8","hashOfConfig":"7"},{"size":748,"mtime":1616857057866,"results":"9","hashOfConfig":"7"},{"size":1661,"mtime":1618066199700,"results":"10","hashOfConfig":"7"},{"size":2656,"mtime":1618066280200,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1ser5sa",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\github\\budget-account-js\\src\\index.js",[],"D:\\github\\budget-account-js\\src\\App.js",[],"D:\\github\\budget-account-js\\src\\views\\Landing.js",[],"D:\\github\\budget-account-js\\src\\views\\Account.js",[],"D:\\github\\budget-account-js\\src\\views\\components\\AccountTable.js",[]]
