+++ src/views/components/AccountTable.js
@@ -1,4 +1,11 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
+
+// 추후 개선방향
+// Account 컴포넌트에서 수입/지출에 대한 데이터를 모두 관리
+// Table 컴포넌트는 데이터 표시만을 담당 (input field 변경은 Account 컴포넌트 state에 직접 반영됨)
+// Account 컴포넌트의 데이터를 Table 컴포넌트에서 직접 변경할 때 객체 구조 등 고민 필요
+
+// 저장 버튼을 눌렀을 때는 localStorage에 해당 월이 key로, 수입 지출이 합쳐진 데이터가 value로 입력되어야 할 듯
 
 const ROW_DATA = {
   date: { type: 'date', value: '' },
@@ -8,21 +15,10 @@ const ROW_DATA = {
   memo: { type: 'text', value: '' },
 };
 
-const AccountTable = ({ title, columns }) => {
-  const [columnNames, setColumnNames] = useState(null);
+const AccountTable = ({ columns }) => {
   const [accountDetail, setAccountDetail] = useState([ROW_DATA]);
 
-  const localStorageKey = title === '수입' ? 'income' : 'expense';
-
-  useEffect(() => {
-    loadAccount();
-  }, []);
-
-  useEffect(() => {
-    if (!columns) return;
-
-    setColumnNames(Object.values(columns));
-  }, [columns]);
+  const columnNames = Object.values(columns).map(({ label }) => label);
 
   const updateAccountDetail = (e, rowData, rowIndex, columnName) => {
     const updatedDetail = accountDetail.map((detail, index) =>
@@ -40,19 +36,6 @@ const AccountTable = ({ title, columns }) => {
     setAccountDetail(accountDetailExceptTargetRow);
   };
 
-  const loadAccount = () => {
-    const localStorage = window.localStorage;
-    const previousData = localStorage.getItem(localStorageKey);
-
-    if (!previousData) return;
-    setAccountDetail(JSON.parse(previousData));
-  };
-
-  const saveAccount = () => {
-    const localStorage = window.localStorage;
-    localStorage.setItem(localStorageKey, JSON.stringify(accountDetail));
-  };
-
   return (
     <>
       <table>
@@ -63,27 +46,27 @@ const AccountTable = ({ title, columns }) => {
           </tr>
         </thead>
         <tbody>
-          {accountDetail.map((row, rowIndex) => (
-            <tr key={`${rowIndex}`}>
-              {Object.entries(row).map(([column, { type, value }], index) => (
-                <td key={`${index}-${column}`}>
-                  <input
-                    {...{ type, value }}
-                    onChange={(e) => updateAccountDetail(e, row, rowIndex, column)}
-                  />
-                </td>
-              ))}
-              {rowIndex !== 0 && (
-                <td>
-                  <button onClick={() => deleteAccountDetail(rowIndex)}>X</button>
-                </td>
-              )}
-            </tr>
-          ))}
+          {accountDetail &&
+            accountDetail.map((row, rowIndex) => (
+              <tr key={`${rowIndex}`}>
+                {Object.entries(row).map(([column, { type, value }], index) => (
+                  <td key={`${index}-${column}`}>
+                    <input
+                      {...{ type, value }}
+                      onChange={(e) => updateAccountDetail(e, row, rowIndex, column)}
+                    />
+                  </td>
+                ))}
+                {rowIndex !== 0 && (
+                  <td>
+                    <button onClick={() => deleteAccountDetail(rowIndex)}>X</button>
+                  </td>
+                )}
+              </tr>
+            ))}
         </tbody>
       </table>
       <button onClick={() => setAccountDetail([...accountDetail, ROW_DATA])}>추가</button>
-      <button onClick={saveAccount}>저장</button>
     </>
   );
 };
